# docs/contracts/openapi.json
{
  "components": {
    "schemas": {
      "ConversationMessageRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "requirement_input",
              "requirement_clarification",
              "status"
            ],
            "title": "Kind",
            "type": "string"
          },
          "message_id": {
            "title": "Message Id",
            "type": "string"
          },
          "payload": {
            "additionalProperties": true,
            "title": "Payload",
            "type": "object"
          },
          "role": {
            "enum": [
              "user",
              "assistant",
              "system"
            ],
            "title": "Role",
            "type": "string"
          },
          "routebook_id": {
            "format": "uuid",
            "title": "Routebook Id",
            "type": "string"
          },
          "workflow_run_id": {
            "format": "uuid",
            "title": "Workflow Run Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "routebook_id",
          "workflow_run_id",
          "message_id",
          "role",
          "kind",
          "payload",
          "created_at"
        ],
        "title": "ConversationMessageRead",
        "type": "object"
      },
      "CreateRouteBookRequest": {
        "additionalProperties": false,
        "properties": {
          "title": {
            "default": "未命名路书",
            "maxLength": 100,
            "minLength": 1,
            "title": "Title",
            "type": "string"
          }
        },
        "title": "CreateRouteBookRequest",
        "type": "object"
      },
...
      "ErrorResponse": {
        "additionalProperties": false,
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorItem"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RouteBookCreationAccepted": {
        "additionalProperties": false,
        "properties": {
          "events_url": {
            "title": "Events Url",
            "type": "string"
          },
          "routebook_id": {
            "format": "uuid",
            "title": "Routebook Id",
            "type": "string"
          },
          "routebook_status": {
            "$ref": "#/components/schemas/RouteBookStatus"
          },
          "status_url": {
            "title": "Status Url",
            "type": "string"
          },
          "workflow_run_id": {
            "format": "uuid",
            "title": "Workflow Run Id",
            "type": "string"
          },
          "workflow_status": {
            "$ref": "#/components/schemas/WorkflowStatus"
          }
        },
        "required": [
          "routebook_id",
          "workflow_run_id",
          "routebook_status",
          "workflow_status",
          "status_url",
          "events_url"
        ],
        "title": "RouteBookCreationAccepted",
        "type": "object"
      },
      "RouteBookMessageCreate": {
        "additionalProperties": false,
        "properties": {
          "message_id": {
            "maxLength": 128,
            "minLength": 8,
            "title": "Message Id",
            "type": "string"
          },
          "text": {
            "maxLength": 4000,
            "minLength": 1,
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "message_id",
          "text"
        ],
        "title": "RouteBookMessageCreate",
        "type": "object"
      },
      "RouteBookRead": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "current_version": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RouteBookVersionRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "current_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Version Id"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "latest_final_version_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Final Version Id"
          },
          "status": {
            "$ref": "#/components/schemas/RouteBookStatus"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status",
          "current_version_id",
          "latest_final_version_id",
          "current_version",
          "created_at",
          "updated_at"
        ],
        "title": "RouteBookRead",
        "type": "object"
      },
...
      "RouteBookStatus": {
        "enum": [
          "draft",
          "planning",
          "pending_confirmation",
          "editable",
          "blocked",
          "final"
        ],
        "title": "RouteBookStatus",
        "type": "string"
      },
...
      "WorkflowStatus": {
        "enum": [
          "queued",
          "running",
          "interrupted",
          "completed",
          "failed",
          "cancelled"
        ],
        "title": "WorkflowStatus",
        "type": "string"
      }
    }
  },
  "info": {
    "title": "RouteBook Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/routebooks": {
      "post": {
        "operationId": "create_routebook_api_routebooks_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRouteBookRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteBookCreationAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Service Unavailable"
          }
        },
        "summary": "Create Routebook",
        "tags": [
          "routebooks"
        ]
      }
    },
    "/api/routebooks/{routebook_id}": {
      "get": {
        "operationId": "get_routebook_api_routebooks__routebook_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "routebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Routebook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteBookRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Routebook",
        "tags": [
          "routebooks"
        ]
      }
    },
    "/api/routebooks/{routebook_id}/messages": {
      "get": {
        "operationId": "get_routebook_messages_api_routebooks__routebook_id__messages_get",
        "parameters": [
          {
            "in": "path",
            "name": "routebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Routebook Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ConversationMessageRead"
                  },
                  "title": "Response Get Routebook Messages Api Routebooks  Routebook Id  Messages Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Routebook Messages",
        "tags": [
          "messages"
        ]
      },
      "post": {
        "operationId": "create_routebook_message_api_routebooks__routebook_id__messages_post",
        "parameters": [
          {
            "in": "path",
            "name": "routebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Routebook Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RouteBookMessageCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementWorkflowAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Routebook Message",
        "tags": [
          "messages"
        ]
      }
    },
    "/api/routebooks/{routebook_id}/versions/{version_id}": {
      "get": {
        "operationId": "get_version_api_routebooks__routebook_id__versions__version_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "routebook_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Routebook Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "version_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Version Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteBookVersionRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Version",
        "tags": [
          "routebooks"
        ]
      }
    },
    "/api/workflow-runs/{run_id}": {
      "get": {
        "operationId": "get_workflow_run_api_workflow_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkflowRunRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Workflow Run",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflow-runs/{run_id}/events": {
      "get": {
        "operationId": "workflow_events_api_workflow_runs__run_id__events_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Workflow Events",
        "tags": [
          "workflows"
        ]
      }
    },
    "/api/workflow-runs/{run_id}/resume": {
      "post": {
        "operationId": "resume_requirement_workflow_api_workflow_runs__run_id__resume_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequirementResumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequirementWorkflowAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Resume Requirement Workflow",
        "tags": [
          "workflows"
        ]
      }
    },
...
  }
}